+++ app/Exceptions/Handler.php
@@ -3,6 +3,7 @@
 namespace Kami\Cocktail\Exceptions;
 
 use Throwable;
+use Sentry\Laravel\Integration;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -47,7 +48,7 @@ class Handler extends ExceptionHandler
     public function register()
     {
         $this->reportable(function (Throwable $e) {
-            //
+            Integration::captureUnhandledException($e);
         });
 
         $this->renderable(function (HttpException $e, $request) {