+++ database/migrations/2024_12_23_215228_add_cocktail_matching_indexes.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+return new class () extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('cocktail_ingredient_substitutes', function (Blueprint $table) {
+            $table->index(['cocktail_ingredient_id', 'ingredient_id'], 'idx_ing_subs_composite');
+        });
+        Schema::table('cocktail_ingredients', function (Blueprint $table) {
+            $table->index(['cocktail_id', 'optional', 'ingredient_id'], 'idx_cocktail_ing_composite');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::table('cocktail_ingredient_substitutes', function (Blueprint $table) {
+            $table->dropIndex('idx_ing_subs_composite');
+        });
+        Schema::table('cocktail_ingredients', function (Blueprint $table) {
+            $table->dropIndex('idx_cocktail_ing_composite');
+        });
+    }
+};